Fix fmap-duplicate performance issue
The fmap_duplicate function is used create a new fmap with a view into
an existing fmap. When the new view is a different size than the old
fmap, a new hash must be calculated for the duplicate fmap. However,
when the duplicated fmap is the same size as the original fmap, the hash
will be the same and there's no point recalculating.

The issue is apparent when scanning large EXE files because the hash was
being calculated at the beginning and end of the scan.

Digging into this issue revealed that hash calculations for fmaps were
also being performed at the wrong place. For scans of maps we use
fmap_duplicate() early in the process to apply the name API argument to
the duplicate fmap. Fixing the logic so we doing recalculate the hash
revealed that we never calculated hashes for fmap's created from buffers
in the first place, so that also had to be fixed be relocating where the
hash is calculated.

I also found that fmap_duplicate()'s offset argument used an off_t,
though it and all caller offsets are not allowed to be negative. This
was a bit of tangent to fix a bunch of off_t variables and paramters
that should've been size_t.

Added a couple unit tests to verify that making duplicate fmaps, and
duplicate-duplicate fmaps works as expected after the change.

Changed CLI_ISCONTAINED() and CLI_ISCONTAINED2() macros to cast to
size_t, because pointers and buffer sizes may not be negative, and these
two macros do not rely on substraction.

+++ libclamav/cpio.c
@@ -103,7 +103,7 @@ int cli_scancpio_old(cli_ctx *ctx)
     unsigned int file = 0, trailer = 0;
     uint32_t filesize, namesize, hdr_namesize;
     int ret         = CL_CLEAN, conv;
-    off_t pos       = 0;
+    size_t pos      = 0;
     int virus_found = 0;
 
     memset(name, 0, sizeof(name));
@@ -195,7 +195,7 @@ int cli_scancpio_odc(cli_ctx *ctx)
     unsigned int file = 0, trailer = 0;
     uint32_t filesize, namesize, hdr_namesize;
     int ret         = CL_CLEAN;
-    off_t pos       = 0;
+    size_t pos      = 0;
     int virus_found = 0;
 
     while (fmap_readn(*ctx->fmap, &hdr_odc, pos, sizeof(hdr_odc)) == sizeof(hdr_odc)) {
@@ -281,7 +281,7 @@ int cli_scancpio_newc(cli_ctx *ctx, int crc)
     unsigned int file = 0, trailer = 0;
     uint32_t filesize, namesize, hdr_namesize, pad;
     int ret         = CL_CLEAN;
-    off_t pos       = 0;
+    size_t pos      = 0;
     int virus_found = 0;
 
     memset(name, 0, 513);
