+++ libclamav/png.c
@@ -426,7 +426,7 @@ scan_overlay:
     /* Check if there's an overlay, and scan it if one exists. */
     if (map->len - offset > 0) {
         cli_dbgmsg("PNG: Found " STDu64 " additional data after end of PNG! Scanning as a nested file.\n", map->len - offset);
-        status = cli_magic_scan_nested_fmap_type(map, offset, map->len - offset, ctx, CL_TYPE_ANY, NULL);
+        status = cli_magic_scan_nested_fmap_type(map, (size_t)offset, map->len - offset, ctx, CL_TYPE_ANY, NULL);
         goto done;
     }
